+++ ValheimServersideQoL/ModConfig.PlayersConfig.cs
@@ -48,6 +48,13 @@ partial record ModConfigBase
         public ConfigEntry<int> InitialBackpackSlots { get; } = cfg.BindEx(section, 4, "Initial available slots in the backpack");
         public ConfigEntry<int> AdditionalBackpackSlotsPerDefeatedBoss { get; } = cfg.BindEx(section, 4, "Additional backpack slots per defeated boss");
         public ConfigEntry<int> MaxBackpackWeight { get; } = cfg.BindEx(section, 0, "Maximum backpack weight. 0 for no limit.");
+        public ConfigEntry<BackPackOnDeathOptions> BackpackOnDeath { get; } = cfg.BindEx(section, BackPackOnDeathOptions.Keep, "What happens to backpack contents on player death");
+
+        public enum BackPackOnDeathOptions
+        {
+            Keep,
+            Destroy
+        }
 
         public ConfigEntry<bool> CanSacrificeMegingjord { get; } = cfg.BindEx(section, false,
             "If true, players can permanently unlock increased carrying weight by sacrificing a megingjord in an obliterator");